+++ Assets/NatShare/Changelog.md
@@ -1,3 +1,11 @@
+## NatShare 1.1f1
++ Added `ShareText` function for sharing plain text.
++ Added `ShareMedia` function for sharing media files like videos, images, and so on.
++ Refactored `Share` function to `ShareImage`.
++ Correctly set duration of video file being saved to camera roll on Android.
++ Fix creation date on saved video being incorrect on Android.
++ Fix sharing not being available for Instagram and Twitter on iOS.
+
 ## NatShare 1.0f3
 + Images being shared or saved will not appear in a "NatShare" album.
 + Added null checking in all functions to prevent crashing.